\name{Nevins}
\alias{Nevins}
\docType{data}
\title{Intensity data for 46 Affymetrix slides with tissue samples of breast tumors}
\description{Intensity data for 46 Affymetrix hu6800 slides with tissue samples of
  breast tumors. See vignette Nevins.pdf in the /scripts directory for details of the
  processing.
}
\usage{data(Nevins)}
\format{\code{Nevins} is an \code{ExpressionSet} containing the data from 46 Affymetrix chips.}
\references{Predicting the clinical status of human breast cancer by using gene expression profiles,
West M, Blanchette C, Dressman H, Huang E, Ishida S, Spang R, Zuzan H,
Olson JA Jr, Marks JR, and Nevins JR. Proc Natl Acad Sci U S A 98(20):11462-7 (2001)}
\source{\url{http://data.cgt.duke.edu/west.php}}
\examples{
data(Nevins)
Nevins
}
\keyword{datasets}
